Create 25.Tuples
Python tuples are very much like python lists, except they are immutable.Python lists are a series of objects separated by commas.The list is often placed inside parentheses, but this is not always necessary.
Singleton lists are ugly, since they must be written with an extra comma. This is because there is no way to tell that single expression is a tuple rather than a plain value, even if it is surrounded by parentheses. For instance, 3 is just the number three. You cannot make a list out of it by writing (3), because that is still a valid way to write the number three. Therefore Python forces you to write the singleton list containing just 3 as (3,).