Create Armstrong
What is an Armstrong number? If the sum of the number of their own digits to the power of the number of digits is equal to that number then that number is called as Armstrong number.

For example : let number be 370 then 370 is of 3 digit number. so if 3^3 + 7^3 + 0^3 = 370 then the number is called as Armstrong number. All one digit number are Armstrong number i.e 1,2,3,4,5,6,7,8,9. Their exist no 2 digit number which is Armstrong number.

For 3 digit 153, 370, 371 and 407 and For 4 digit 1634, 8208 and 9474 are the example of Armstrong number.