Create Factorial
For an integer n greater than or equal to 1, the factorial is the product of all integers less than or equal to n but greater than or equal to 1. The factorial value of 0 is defined as equal to 1. The factorial values for negative integers are not defined.

For eg : factorial of 5 i.e 5!=5*4*3*2*1=120.

While calculating permutation and combination, Factorial program can be referred.